requirements.txt
tensorflow-cpu==2.20.0
keras==2.15.0
streamlit==1.39.0
opencv-python-headless
numpy
pillow
